Fix (message) : Id 타입 Long으로 통일
- 계속되었던 타입 오류를 고쳤다.

related to #2

+++ src/main/java/com/rolling/domain/message/MessageRepository.java
@@ -3,6 +3,6 @@ package com.rolling.domain.message;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
-public interface MessageRepository extends JpaRepository<Message, Integer> {
+public interface MessageRepository extends JpaRepository<Message, Long> {
 
 }
